requirements.txt
click~=7.1.2
colorama~=0.4.4
dominate~=2.6.0
Flask~=1.1.2
Flask-Bootstrap4~=4.0.2
Flask-WTF~=0.15.1
itsdangerous~=1.1.0
Jinja2~=2.11.3
MarkupSafe~=1.1.1
visitor~=0.1.3
Werkzeug~=1.0.1
WTForms~=2.3.3
